refactor(NameTest): 테스트 메서드를 더 적합한 테스트 클래스로 이동
