+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -5,7 +5,6 @@ import static blackjack.domain.card.CardNumber.KING;
 import static blackjack.domain.card.CardNumber.QUEEN;
 import static blackjack.domain.card.CardSymbol.DIAMOND;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
@@ -16,32 +15,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.EmptySource;
 
 class PlayerTest {
-
-    @ParameterizedTest
-    @EmptySource
-    @DisplayName("이름이 유효하지 않으면 예외를 던진다.")
-    void emptyName(String name) {
-        // then
-        assertThatThrownBy(() -> new Player(name))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("공백은 허용되지 않습니다.");
-    }
-
-    @Test
-    @DisplayName("이름이 제한된 길이를 초과하면 예외를 던진다.")
-    void nameLength() {
-        // give
-        final String name = "1234567890".repeat(10) + "1";
-
-        // then
-        assertThatThrownBy(() -> new Player(name))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("길이는 100자를 초과할 수 없습니다.");
-
-    }
 
     @ParameterizedTest
     @CsvSource(value = {"TWO:false", "ACE:true"}, delimiter = ':')
